+++ .gitignore
@@ -1,3 +1,5 @@
 node_modules
+tsconfig.tsbuildinfo
+
 dist
-incomplete
+scripts
