+++ emily.md/index.md
@@ -1,0 +1,9 @@
+# emily
+
+hello this is me. I go by emilywithoutorgans online.
+
+technically my name is émilie, but you can call me either em, emmy, or emily.
+
+<!-- [my webcomic, p~~over~~ty](https://pty.emily.md/) -->
+
+[my notes on category theory](https://notes.emily.md/)
