+++ emily.md/tsconfig.json
@@ -1,0 +1,9 @@
+{
+    "extends": "../tsconfig.base.json",
+    "compilerOptions": {
+        "composite": true,
+        "declaration": true,
+        "outDir": "scripts"
+    },
+    "include": ["build.ts"]
+}
