--- main_index.html
@@ -1,5 +1,0 @@
-<h1>emily</h1>
-<p>hello this is me. I go by emilywithoutorgans online.</p>
-<p>technically my name is émilie, but you can call me either em, emmy, or emily.</p>
-<!-- <p><a href="https://pty.emily.md/">my webcomic, p<s>over</s>ty</a></p> -->
-<p><a href="https://notes.emily.md/">my notes on category theory</a></p>
