+++ markdown-renderer/bottom.css
@@ -1,43 +1,3 @@
-:root {
-    max-width: 70ch;
-}
-
-h1:first-child {
-    margin: 0.67em 0;
-}
-
-
-h1,
-h2,
-h3,
-h4 {
-    margin: 1.5em 0 0.5em 0;
-}
-
-h1 {
-    font-size: 1.875rem;
-}
-
-h2 {
-    font-size: 1.625rem;
-}
-
-h3 {
-    font-size: 1.375rem;
-}
-
-h4 {
-    font-size: 1.25rem;
-}
-
-p {
-    font-size: 1rem;
-    margin: 0 0 1rem 0;
-}
-
-small {
-    font-size: 0.889rem;
-}
 
 blockquote {
     font-style: italic;
@@ -49,15 +9,6 @@ blockquote {
 
 a {
     color: #007793;
-}
-
-iframe {
-    width: 100%;
-    margin-top: -30px;
-    margin-bottom: -10px;
-    position: relative;
-    z-index: -1;
-    user-select: none;
 }
 
 .hljs-strong {
