+++ markdown-renderer/package.json
@@ -1,0 +1,19 @@
+{
+    "name": "@emily/markdown-renderer",
+    "private": true,
+    "author": "emilywithoutorgans",
+    "license": "MIT",
+    "type": "module",
+    "main": "dist/index.js",
+    "types": "dist/index.d.ts",
+    "dependencies": {
+        "csso": "^5.0.5",
+        "highlight.js": "^11.11.1",
+        "marked": "^15.0.4",
+        "marked-highlight": "^2.2.1"
+    },
+    "devDependencies": {
+        "@types/csso": "^5.0.4",
+        "@types/marked": "^5.0.0"
+    }
+}
