+++ markdown-renderer/tsconfig.json
@@ -1,0 +1,10 @@
+{
+    "extends": "../tsconfig.base.json",
+    "compilerOptions": {
+        "outDir": "dist",
+        "rootDir": "src",
+        "composite": true,
+        "declaration": true
+    },
+    "include": ["src"]
+}
