+++ notes.emily.md/build.ts
@@ -1,0 +1,26 @@
+import fs from "fs/promises";
+import path from "path";
+import { renderMarkdownToHTML } from "@emily/markdown-renderer";
+
+const distDir = path.resolve("dist");
+await fs.mkdir(distDir, { recursive: true });
+
+let indexMarkdown = `# notes`;
+
+const markdownDir = path.resolve("markdown");
+for (const file of await fs.readdir(markdownDir)) {
+    if (path.extname(file) !== ".md") continue;
+
+    const filePath = path.join(markdownDir, file);
+    const outputFileName = `${path.basename(file, ".md")}.html`;
+    const outputFilePath = path.join(distDir, outputFileName + ".html"); // double .html to fix route matching
+
+    const markdown = await fs.readFile(filePath, "utf-8");
+    await fs.writeFile(outputFilePath, renderMarkdownToHTML(markdown));
+
+    indexMarkdown += `\n[/${outputFileName}](/${outputFileName})`;
+
+    console.log(`${file} -> ${path.relative(process.cwd(), outputFilePath)}`);
+}
+
+await fs.writeFile(path.join(distDir, "index.html"), renderMarkdownToHTML(indexMarkdown));
