+++ notes.emily.md/markdown/1844-manuscripts.md
