+++ notes.emily.md/markdown/simplicial-sets.md
